requirements.txt
fastapi
uvicorn
python-multipart
torch
torchaudio
librosa
soundfile
numpy
gTTS
pydub
scikit-learn
torchvision
python-dotenv
